Refactored and tested

+++ Backend/cy-finance/src/main/java/com/example/demo/users/UserController.java
@@ -7,19 +7,20 @@ import com.example.demo.expenses.ExpensesRepository;
 import com.example.demo.netWorth.NetWorth;
 import com.example.demo.netWorth.NetWorthRepository;
 import com.example.demo.util.Response;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Example;
 import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 public class UserController {
 
     Response<User> users = new Response<>();
+
+    private final Logger logger = LoggerFactory.getLogger(UserController.class);
 
     @Autowired
     EarningsRepository earningsRepository;
@@ -35,64 +36,73 @@ public class UserController {
 
     @GetMapping("/users")
     public ResponseEntity<List<User>> getUsers(
-            @CookieValue(name = "user-id", required = false) String userId)
-    {
-        if (userId == null
-                || userId.isEmpty()
-                || !userRepository.findById(Integer.parseInt(userId)).getRole().equals("admin"))
-        {
-            return ResponseEntity.ok(null);
-        } else {
+            @CookieValue(name = "user-id", required = false) String userId) {
+        logger.info("[GET /users] Cookie: " + userId);
+        if (isValidUserId(userId) && isAdmin(userId)) {
+            logger.info("[GET /users] Successfully accessed data by user: " + userRepository.findById(Integer.parseInt(userId)).getName());
             return ResponseEntity.ok(userRepository.findAll());
+        } else {
+            if (isValidUserId(userId))
+                logger.info("[GET /users] Role: " + userRepository.findById(Integer.parseInt(userId)).getRole());
+            logger.warn("[GET /users] Attempted access from invalid user");
+            return ResponseEntity.ok(null);
         }
     }
 
     @GetMapping("/users/{id}")
     public ResponseEntity<User> getUser(@PathVariable int id,
-                                        @CookieValue(name = "user-id", required = false) String userId){
+                                        @CookieValue(name = "user-id", required = false) String userId) {
+        logger.info("[GET /users/{id}] Cookie: " + userId);
         // TODO: Find out how to make sure the empty cookie request is only for signup
         // Only return user if the cookie is set and the user is either an admin or the requested user
-        if (userId == null
-                || userId.isEmpty()
-                || (!userRepository.findById(Integer.parseInt(userId)).getRole().equals("admin") && id != Integer.parseInt(userId))) {
-            return ResponseEntity.ok(null);
-        } else {
+        if (isValidUserId(userId) && (isAdmin(userId) || id == Integer.parseInt(userId))) {
+            logger.info("[GET /users/{id}] Successfully accessed data by user: " + userRepository.findById(Integer.parseInt(userId)).getName());
             return ResponseEntity.ok(userRepository.findById(id));
+        } else {
+            logger.warn("[GET /users/{id}] Attempted access from invalid user");
+            return ResponseEntity.ok(null);
         }
     }
 
     @PostMapping("/users")
     public ResponseEntity<Response<String>> setUser(@RequestBody User user,
-                                                    @CookieValue(name = "user-id", required = false) String userId){
+                                                    @CookieValue(name = "user-id", required = false) String userId) {
+        logger.info("[POST /users] Cookie: " + userId);
+        logger.info("[POST /users] User: " + user);
         Response<String> response = new Response<>();
         // Able to create user if not logged in (for signup) or admin
-        if (userId == null || userRepository.findById(Integer.parseInt(userId)).getRole().equals("admin")) {
+        if (userId == null || (isValidUserId(userId) && isAdmin(userId))) {
             if (user == null) {
-                response.put("message", "No user specified");
-                return ResponseEntity.ok(response);
+                logger.warn("[POST /users] No user provided");
+                response.put("message", "No user provided");
             } else {
                 userRepository.save(user);
+                logger.info("[POST /users] User created: " + user);
                 response.put("message", "User created");
-                return ResponseEntity.ok(response);
             }
         } else {
+            logger.warn("[POST /users] Attempted access from invalid user");
             response.put("message", "User creation not allowed");
-            return ResponseEntity.ok(response);
         }
+        return ResponseEntity.ok(response);
     }
 
     @PostMapping("/login")
-    public ResponseEntity<Response<String>> login(@RequestParam String email, @RequestParam String password){
+    public ResponseEntity<Response<String>> login(@RequestParam String email, @RequestParam String password) {
         Response<String> response = new Response<>();
 
         User foundUser = userRepository.findByEmail(email);
-        if (foundUser.getPassword().equals(password)){
-            response.put("message", "success");
+        if (foundUser.getPassword().equals(password)) {
             ResponseCookie springCookie = ResponseCookie.from("user-id", String.valueOf(foundUser.getId()))
                     .maxAge(60)
                     .build();
+            HttpHeaders responseHeaders = new HttpHeaders();
+            responseHeaders.set(HttpHeaders.SET_COOKIE, springCookie.toString());
+
+            response.put("message", "success");
+
             return ResponseEntity.ok()
-                    .header(HttpHeaders.SET_COOKIE, springCookie.toString())
+                    .headers(responseHeaders)
                     .body(response);
         } else {
             response.put("message", "failure");
@@ -101,7 +111,7 @@ public class UserController {
     }
 
     @PostMapping("/logout")
-    public ResponseEntity<Response<String>> logout(){
+    public ResponseEntity<Response<String>> logout() {
         Response<String> response = new Response<>();
         response.put("message", "Successfully logged out");
         ResponseCookie springCookie = ResponseCookie
@@ -114,33 +124,33 @@ public class UserController {
 
     @PutMapping("/users")
     public ResponseEntity<Response<String>> changeUser(@RequestBody User user,
-                                                       @CookieValue(name = "user-id", required = false) String userId){
+                                                       @CookieValue(name = "user-id", required = false) String userId) {
         Response<String> response = new Response<>();
 
-        if (user == null) {
-            response.put("message", "No user provided");
-            return ResponseEntity.ok(response);
-
-            // Only edit user if the cookie is set and the user is either an admin or the requested user
-        } else if (userId.isEmpty()
-                || (!userRepository.findById(Integer.parseInt(userId)).getRole().equals("admin")
-                    && user.getId() != Integer.parseInt(userId))) {
-
-            response.put("message", "User not allowed to perform this action");
-            return ResponseEntity.ok(response);
+        logger.info("[PUT /users] Cookie: " + userId);
+        // Only edit user if the cookie is set and the user is either an admin or the requested user
+        if (isValidUserId(userId) && (isAdmin(userId) || user.getId() == Integer.parseInt(userId))) {
+            if (user == null) {
+                logger.warn("[PUT /users] User not provided");
+                response.put("message", "No user provided");
+            } else {
+                userRepository.save(user);
+                logger.info("[PUT /users] User " + user.getName() + " modified by " + userRepository.findById(Integer.parseInt(userId)).getEmail());
+                response.put("message", "User modified");
+            }
         } else {
-            userRepository.save(user);
-            response.put("message", "User modified");
-            return ResponseEntity.ok(response);
+            logger.warn("[PUT /users] Attempted access from invalid user");
+            response.put("message", "User not allowed to perform this action");
         }
+        return ResponseEntity.ok(response);
     }
 
     @PostMapping("/users/{userId}/earnings/{earningsId}")
-    String assignEarningsToUser(@PathVariable int userId,@PathVariable int earningsId){
+    String assignEarningsToUser(@PathVariable int userId, @PathVariable int earningsId) {
         Response<String> response = new Response<>();
         User user = userRepository.findById(userId);
         Earnings earnings = earningsRepository.findById(earningsId);
-        if(user == null || earnings == null) {
+        if (user == null || earnings == null) {
             response.put("message", "Failed to assign earnings");
         } else {
             earnings.setUser(user);
@@ -152,10 +162,17 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<Response<String>> deleteUser(@PathVariable int id){
+    public ResponseEntity<Response<String>> deleteUser(@PathVariable int id,
+                                                       @CookieValue(name = "user-id", required = false) String userId) {
         Response<String> response = new Response<>();
-        userRepository.deleteById(id);
-        response.put("message", "User deleted");
+        if (isValidUserId(userId) && (isAdmin(userId) || id == Integer.parseInt(userId))) {
+            userRepository.deleteById(id);
+            logger.info("[DELETE /users] Entry for userId " + id + " deleted by userId " + userId);
+            response.put("message", "User deleted");
+        } else {
+            logger.warn("[DELETE /users] Attempted access from invalid user");
+            response.put("message", "User not allowed to perform this action");
+        }
         return ResponseEntity.ok(response);
     }
 
@@ -164,7 +181,7 @@ public class UserController {
         Response<String> response = new Response<>();
         User user = userRepository.findById(userId);
         NetWorth netWorth = netWorthRepository.findById(netWorthId);
-        if(user == null || netWorth == null) {
+        if (user == null || netWorth == null) {
             response.put("message", "Failed to assign net worth");
             return ResponseEntity.ok(response);
         } else {
@@ -181,7 +198,7 @@ public class UserController {
         Response<String> response = new Response<>();
         User user = userRepository.findById(userId);
         Expenses expenses = expensesRepository.findById(expensesId);
-        if(user == null || expenses == null) {
+        if (user == null || expenses == null) {
             response.put("message", "Failed to assign expenses");
         } else {
             expenses.setUser(user);
@@ -190,5 +207,17 @@ public class UserController {
             response.put("message", "Expenses assigned");
         }
         return response.toString();
+    }
+
+    private boolean isValidId(String userId) {
+        return userId != null && !userId.isEmpty();
+    }
+
+    private boolean isValidUserId(String userId) {
+        return isValidId(userId) && userRepository.findById(Integer.parseInt(userId)) != null;
+    }
+
+    private boolean isAdmin(String userId) {
+        return userRepository.findById(Integer.parseInt(userId)).getRole().equals("admin");
     }
 }
